requirements.txt
opencv-python
numpy
streamlit
pyqt5
Pillow